feat(wallet-sdk): own the spark wallet init (connection-core step 4b)
The SDK now owns the spark/Breez connection layer — the last connection piece the
AccountRepository depends on.

- src/spark.ts (split out of features/shared/spark.ts, bodies verbatim): sparkDebugLog,
  the Breez initLogging once-guard, the mnemonic/identity-pubkey/wallet queryOptions
  (converted to plain query-options objects) and getInitializedSparkWallet (offline stub
  handling). The two web bindings are replaced by a configureSpark({apiKey,
  isDebugLoggingEnabled}) seam: the web supplies the env-derived Breez API key and the
  per-call DEBUG_LOGGING_SPARK feature-flag check; log formats unchanged.
- src/spark-utils.ts (moved verbatim from lib/spark/utils.ts): the Breez external signer
  identity-pubkey derivation + the offline wallet stub. lib/spark keeps errors.ts + wasm.ts
  (their consumers move in later phases) and re-exports the moved utils.
- web's features/shared/spark.ts keeps what is web: the eager VITE_BREEZ_API_KEY read/throw,
  the configureSpark call (module load — same ordering + fail-fast as today, and the
  server-side lightning-address path configures identically since it imports via this module),
  and the useTrackAndUpdateSparkAccountBalances React hook. Everything else re-exported (all
  consumers untouched).
- sparkDebugLog being SDK-owned dissolves the previously planned AccountsCache logger
  injection: when the cache moves (next chunk), it imports the SDK logger directly.

typecheck, biome, build all green; 138 tests; no new SDK deps.
