+++ apps/web-wallet/app/lib/spark/index.ts
@@ -1,3 +1,5 @@
 export * from './errors';
-export * from './utils';
+// utils moved to @agicash/wallet-sdk (spark-utils); re-exported here until the
+// import-cleanup PR.
+export * from '@agicash/wallet-sdk/spark-utils';
 export * from './wasm';
