+++ packages/wallet-sdk/src/spark-utils.ts
